runtime: Create common folder for v0.0.5 wasm_tests
