+++ runtime/test/wasm_test/api_version_0_0_5/abi_classes.ts
@@ -1,31 +1,5 @@
-enum IndexForAscTypeId {
-  STRING = 0,
-  ARRAY_BUFFER = 1,
-  UINT8_ARRAY = 6,
-  ARRAY_STRING = 18,
-}
-
-export function id_of_type(type_id_index: IndexForAscTypeId): usize {
-  switch (type_id_index) {
-    case IndexForAscTypeId.STRING:
-      return idof<string>();
-    case IndexForAscTypeId.ARRAY_BUFFER:
-      return idof<ArrayBuffer>();
-    case IndexForAscTypeId.UINT8_ARRAY:
-      return idof<Uint8Array>();
-    case IndexForAscTypeId.ARRAY_STRING:
-      return idof<Array<string>>();
-    default:
-      return 0;
-  }
-}
-
-export function allocate(n: usize): usize {
-  return __alloc(n);
-}
-
-// Sequence of 20 `u8`s.
-type Address = Uint8Array;
+export * from './common/global'
+import { Address, Uint8, FixedBytes, Bytes, Payload, Value } from './common/types'
 
 // Clone the address to a new buffer, add 1 to the first and last bytes of the
 // address and return the new address.
@@ -36,15 +10,12 @@ export function test_address(address: Address): Address {
   new_address[0] += 1;
   new_address[address.length - 1] += 1;
 
-  return new_address
+  return changetype<Address>(new_address)
 }
 
-// Sequence of 32 `u8`s.
-type Uint = Uint8Array;
-
-// Clone the Uint to a new buffer, add 1 to the first and last `u8`s and return
-// the new Uint.
-export function test_uint(address: Uint): Uint {
+// Clone the Uint8 to a new buffer, add 1 to the first and last `u8`s and return
+// the new Uint8
+export function test_uint(address: Uint8): Uint8 {
   let new_address = address.subarray();
 
   // Add 1 to the first byte.
@@ -57,10 +28,6 @@ export function test_uint(address: Uint): Uint {
 export function repeat_twice(original: string): string {
   return original.repeat(2)
 }
-
-// Sequences of `u8`s.
-type FixedBytes = Uint8Array;
-type Bytes = Uint8Array;
 
 // Concatenate two byte sequences into a new one.
 export function concat(bytes1: Bytes, bytes2: FixedBytes): Bytes {
@@ -82,25 +49,7 @@ export function concat(bytes1: Bytes, bytes2: FixedBytes): Bytes {
 
   let new_typed_array = Uint8Array.wrap(concated_buff);
 
-  return new_typed_array;
-}
-
-enum ValueKind {
-    STRING = 0,
-    INT = 1,
-    BIG_DECIMAL = 2,
-    BOOL = 3,
-    ARRAY = 4,
-    NULL = 5,
-    BYTES = 6,
-    BIG_INT = 7,
-}
-
-// Big enough to fit any pointer or native `this.data`.
-type Payload = u64
-export class Value {
-    kind: ValueKind
-    data: Payload
+  return changetype<Bytes>(new_typed_array);
 }
 
 export function test_array(strings: Array<string>): Array<string> {
