+++ runtime/test/wasm_test/api_version_0_0_5/abi_store_value.ts
@@ -1,61 +1,5 @@
-enum IndexForAscTypeId {
-  STRING = 0,
-  ARRAY_BUFFER = 1,
-  UINT8_ARRAY = 6,
-  BIG_DECIMAL = 12,
-  ARRAY_STRING = 18,
-  STORE_VALUE = 31,
-}
-
-export function id_of_type(type_id_index: IndexForAscTypeId): usize {
-  switch (type_id_index) {
-    case IndexForAscTypeId.STRING:
-      return idof<string>();
-    case IndexForAscTypeId.ARRAY_BUFFER:
-      return idof<ArrayBuffer>();
-    case IndexForAscTypeId.UINT8_ARRAY:
-      return idof<Uint8Array>();
-    case IndexForAscTypeId.BIG_DECIMAL:
-      return idof<BigDecimal>();
-    case IndexForAscTypeId.ARRAY_STRING:
-      return idof<Array<string>>();
-    case IndexForAscTypeId.STORE_VALUE:
-      return idof<Value>();
-    default:
-      return 0;
-  }
-}
-
-export function allocate(n: usize): usize {
-  return __alloc(n);
-}
-
-enum ValueKind {
-    STRING = 0,
-    INT = 1,
-    BIG_DECIMAL = 2,
-    BOOL = 3,
-    ARRAY = 4,
-    NULL = 5,
-    BYTES = 6,
-    BIG_INT = 7,
-}
-
-// Big enough to fit any pointer or native `this.data`.
-type Payload = u64
-
-type Bytes = Uint8Array;
-type BigInt = Uint8Array;
-
-export class BigDecimal {
-    exp!: BigInt
-    digits!: BigInt
-}
-
-export class Value {
-    kind: ValueKind
-    data: Payload
-}
+export * from './common/global'
+import { BigInt, BigDecimal, Bytes, Value, ValueKind } from './common/types'
 
 export function value_from_string(str: string): Value {
     let token = new Value();
