+++ runtime/test/wasm_test/api_version_0_0_5/big_int_arithmetic.ts
@@ -1,27 +1,5 @@
-enum IndexForAscTypeId {
-  STRING = 0,
-  ARRAY_BUFFER = 1,
-  UINT8_ARRAY = 6,
-}
-
-export function id_of_type(type_id_index: IndexForAscTypeId): usize {
-  switch (type_id_index) {
-    case IndexForAscTypeId.STRING:
-      return idof<string>();
-    case IndexForAscTypeId.ARRAY_BUFFER:
-      return idof<ArrayBuffer>();
-    case IndexForAscTypeId.UINT8_ARRAY:
-      return idof<Uint8Array>();
-    default:
-      return 0;
-  }
-}
-
-export function allocate(n: usize): usize {
-  return __alloc(n);
-}
-
-type BigInt = Uint8Array;
+export * from './common/global'
+import { BigInt } from './common/types'
 
 declare namespace bigInt {
     function plus(x: BigInt, y: BigInt): BigInt
