+++ runtime/test/wasm_test/api_version_0_0_5/crypto.ts
@@ -1,25 +1,4 @@
-enum IndexForAscTypeId {
-  STRING = 0,
-  ARRAY_BUFFER = 1,
-  UINT8_ARRAY = 6,
-}
-
-export function id_of_type(type_id_index: IndexForAscTypeId): usize {
-  switch (type_id_index) {
-    case IndexForAscTypeId.STRING:
-      return idof<string>();
-    case IndexForAscTypeId.ARRAY_BUFFER:
-      return idof<ArrayBuffer>();
-    case IndexForAscTypeId.UINT8_ARRAY:
-      return idof<Uint8Array>();
-    default:
-      return 0;
-  }
-}
-
-export function allocate(n: usize): usize {
-  return __alloc(n);
-}
+export * from './common/global'
 
 declare namespace crypto {
     function keccak256(input: Uint8Array): Uint8Array
