+++ runtime/test/wasm_test/api_version_0_0_5/store.ts
@@ -1,58 +1,5 @@
-enum IndexForAscTypeId {
-  STRING = 0,
-  ARRAY_BUFFER = 1,
-  UINT8_ARRAY = 6,
-  BIG_DECIMAL = 12,
-  ARRAY_BOOL = 13,
-  ARRAY_UINT8_ARRAY = 14,
-  ARRAY_STORE_VALUE = 16,
-  ARRAY_STRING = 18,
-  ARRAY_TYPED_MAP_ENTRY_STRING_STORE_VALUE = 21,
-  STORE_VALUE = 31,
-  TYPED_MAP_ENTRY_STRING_STORE_VALUE = 34,
-  TYPED_MAP_STRING_STORE_VALUE = 36,
-  ARRAY_I32 = 47,
-  ARRAY_BIG_DECIMAL = 51,
-}
-
-export function id_of_type(type_id_index: IndexForAscTypeId): usize {
-  switch (type_id_index) {
-    case IndexForAscTypeId.STRING:
-      return idof<string>();
-    case IndexForAscTypeId.ARRAY_BUFFER:
-      return idof<ArrayBuffer>();
-    case IndexForAscTypeId.UINT8_ARRAY:
-      return idof<Uint8Array>();
-    case IndexForAscTypeId.BIG_DECIMAL:
-      return idof<BigDecimal>();
-    case IndexForAscTypeId.ARRAY_BOOL:
-      return idof<Array<bool>>();
-    case IndexForAscTypeId.ARRAY_UINT8_ARRAY:
-      return idof<Array<Uint8Array>>();
-    case IndexForAscTypeId.ARRAY_STORE_VALUE:
-      return idof<Array<Value>>();
-    case IndexForAscTypeId.ARRAY_STRING:
-      return idof<Array<string>>();
-    case IndexForAscTypeId.ARRAY_TYPED_MAP_ENTRY_STRING_STORE_VALUE:
-      return idof<Array<Entity>>();
-    case IndexForAscTypeId.STORE_VALUE:
-      return idof<Value>();
-    case IndexForAscTypeId.TYPED_MAP_ENTRY_STRING_STORE_VALUE:
-      return idof<Entity>();
-    case IndexForAscTypeId.TYPED_MAP_STRING_STORE_VALUE:
-      return idof<TypedMap<string, Value>>();
-    case IndexForAscTypeId.ARRAY_I32:
-      return idof<Array<i32>>();
-    case IndexForAscTypeId.ARRAY_BIG_DECIMAL:
-      return idof<Array<BigDecimal>>();
-    default:
-      return 0;
-  }
-}
-
-export function allocate(n: usize): usize {
-  return __alloc(n);
-}
+export * from './common/global'
+import { TypedMap, Entity, BigDecimal, Value } from './common/types'
 
 /** Definitions copied from graph-ts/index.ts */
 declare namespace store {
@@ -85,478 +32,6 @@ declare namespace bigDecimal {
   function equals(x: BigDecimal, y: BigDecimal): boolean
   function toString(bigDecimal: BigDecimal): string
   function fromString(s: string): BigDecimal
-}
-
-/**
- * TypedMap entry.
- */
-class TypedMapEntry<K, V> {
-  key: K
-  value: V
-
-  constructor(key: K, value: V) {
-    this.key = key
-    this.value = value
-  }
-}
-
-/** Typed map */
-class TypedMap<K, V> {
-  entries: Array<TypedMapEntry<K, V>>
-
-  constructor() {
-    this.entries = new Array<TypedMapEntry<K, V>>(0)
-  }
-
-  set(key: K, value: V): void {
-    let entry = this.getEntry(key)
-    if (entry !== null) {
-      entry.value = value
-    } else {
-      let entry = new TypedMapEntry<K, V>(key, value)
-      this.entries.push(entry)
-    }
-  }
-
-  getEntry(key: K): TypedMapEntry<K, V> | null {
-    for (let i: i32 = 0; i < this.entries.length; i++) {
-      if (this.entries[i].key == key) {
-        return this.entries[i]
-      }
-    }
-    return null
-  }
-
-  get(key: K): V | null {
-    for (let i: i32 = 0; i < this.entries.length; i++) {
-      if (this.entries[i].key == key) {
-        return this.entries[i].value
-      }
-    }
-    return null
-  }
-
-  isSet(key: K): bool {
-    for (let i: i32 = 0; i < this.entries.length; i++) {
-      if (this.entries[i].key == key) {
-        return true
-      }
-    }
-    return false
-  }
-}
-
-/**
- * Byte array
- */
-class ByteArray extends Uint8Array {
-  toHex(): string {
-    return typeConversion.bytesToHex(this)
-  }
-
-  toHexString(): string {
-    return typeConversion.bytesToHex(this)
-  }
-
-  toString(): string {
-    return typeConversion.bytesToString(this)
-  }
-
-  toBase58(): string {
-    return typeConversion.bytesToBase58(this)
-  }
-}
-
-/** A dynamically-sized byte array. */
-class Bytes extends ByteArray { }
-
-/** An Ethereum address (20 bytes). */
-class Address extends Bytes {
-  static fromString(s: string): Address {
-    return typeConversion.stringToH160(s) as Address
-  }
-}
-
-/** An arbitrary size integer represented as an array of bytes. */
-class BigInt extends Uint8Array {
-  toHex(): string {
-    return typeConversion.bigIntToHex(this)
-  }
-
-  toHexString(): string {
-    return typeConversion.bigIntToHex(this)
-  }
-
-  toString(): string {
-    return typeConversion.bigIntToString(this)
-  }
-
-  static fromI32(x: i32): BigInt {
-    return typeConversion.i32ToBigInt(x) as BigInt
-  }
-
-  toI32(): i32 {
-    return typeConversion.bigIntToI32(this)
-  }
-
-  @operator('+')
-  plus(other: BigInt): BigInt {
-    return bigInt.plus(this, other)
-  }
-
-  @operator('-')
-  minus(other: BigInt): BigInt {
-    return bigInt.minus(this, other)
-  }
-
-  @operator('*')
-  times(other: BigInt): BigInt {
-    return bigInt.times(this, other)
-  }
-
-  @operator('/')
-  div(other: BigInt): BigInt {
-    return bigInt.dividedBy(this, other)
-  }
-
-  divDecimal(other: BigDecimal): BigDecimal {
-    return bigInt.dividedByDecimal(this, other)
-  }
-
-  @operator('%')
-  mod(other: BigInt): BigInt {
-    return bigInt.mod(this, other)
-  }
-
-  @operator('==')
-  equals(other: BigInt): boolean {
-    if (this.length !== other.length) {
-      return false;
-    }
-    for (let i = 0; i < this.length; i++) {
-      if (this[i] !== other[i]) {
-        return false;
-      }
-    }
-    return true;
-  }
-
-  toBigDecimal(): BigDecimal {
-    return new BigDecimal(this)
-  }
-}
-
-class BigDecimal {
-  digits: BigInt
-  exp: BigInt
-
-  constructor(bigInt: BigInt) {
-    this.digits = bigInt
-    this.exp = BigInt.fromI32(0)
-  }
-
-  static fromString(s: string): BigDecimal {
-    return bigDecimal.fromString(s)
-  }
-
-  toString(): string {
-    return bigDecimal.toString(this)
-  }
-
-  truncate(decimals: i32): BigDecimal {
-    let digitsRightOfZero = this.digits.toString().length + this.exp.toI32()
-    let newDigitLength = decimals + digitsRightOfZero
-    let truncateLength = this.digits.toString().length - newDigitLength
-    if (truncateLength < 0) {
-      return this
-    } else {
-      for (let i = 0; i < truncateLength; i++) {
-        this.digits = this.digits.div(BigInt.fromI32(10))
-      }
-      this.exp = BigInt.fromI32(decimals* -1)
-      return this
-    }
-  }
-
-  @operator('+')
-  plus(other: BigDecimal): BigDecimal {
-    return bigDecimal.plus(this, other)
-  }
-
-  @operator('-')
-  minus(other: BigDecimal): BigDecimal {
-    return bigDecimal.minus(this, other)
-  }
-
-  @operator('*')
-  times(other: BigDecimal): BigDecimal {
-    return bigDecimal.times(this, other)
-  }
-
-  @operator('/')
-  div(other: BigDecimal): BigDecimal {
-    return bigDecimal.dividedBy(this, other)
-  }
-
-  @operator('==')
-  equals(other: BigDecimal): boolean {
-    return bigDecimal.equals(this, other)
-  }
-}
-
-/**
- * Enum for supported value types.
- */
-enum ValueKind {
-  STRING = 0,
-  INT = 1,
-  BIGDECIMAL = 2,
-  BOOL = 3,
-  ARRAY = 4,
-  NULL = 5,
-  BYTES = 6,
-  BIGINT = 7,
-}
-
-/**
- * Pointer type for Value data.
- *
- * Big enough to fit any pointer or native `this.data`.
- */
-type ValuePayload = u64
-
-/**
- * A dynamically typed value.
- */
-class Value {
-  kind: ValueKind
-  data: ValuePayload
-
-  toAddress(): Address {
-    assert(this.kind == ValueKind.BYTES, 'Value is not an address.')
-    return changetype<Address>(this.data as u32)
-  }
-
-  toBoolean(): boolean {
-    if (this.kind == ValueKind.NULL) {
-      return false;
-    }
-    assert(this.kind == ValueKind.BOOL, 'Value is not a boolean.')
-    return this.data != 0
-  }
-
-  toBytes(): Bytes {
-    assert(this.kind == ValueKind.BYTES, 'Value is not a byte array.')
-    return changetype<Bytes>(this.data as u32)
-  }
-
-  toI32(): i32 {
-    if (this.kind == ValueKind.NULL) {
-      return 0;
-    }
-    assert(this.kind == ValueKind.INT, 'Value is not an i32.')
-    return this.data as i32
-  }
-
-  toString(): string {
-    assert(this.kind == ValueKind.STRING, 'Value is not a string.')
-    return changetype<string>(this.data as u32)
-  }
-
-  toBigInt(): BigInt {
-    assert(this.kind == ValueKind.BIGINT, 'Value is not a BigInt.')
-    return changetype<BigInt>(this.data as u32)
-  }
-
-  toBigDecimal(): BigDecimal {
-    assert(this.kind == ValueKind.BIGDECIMAL, 'Value is not a BigDecimal.')
-    return changetype<BigDecimal>(this.data as u32)
-  }
-
-  toArray(): Array<Value> {
-    assert(this.kind == ValueKind.ARRAY, 'Value is not an array.')
-    return changetype<Array<Value>>(this.data as u32)
-  }
-
-  toBooleanArray(): Array<boolean> {
-    let values = this.toArray()
-    let output = new Array<boolean>(values.length)
-    for (let i: i32; i < values.length; i++) {
-      output[i] = values[i].toBoolean()
-    }
-    return output
-  }
-
-  toBytesArray(): Array<Bytes> {
-    let values = this.toArray()
-    let output = new Array<Bytes>(values.length)
-    for (let i: i32 = 0; i < values.length; i++) {
-      output[i] = values[i].toBytes()
-    }
-    return output
-  }
-
-  toStringArray(): Array<string> {
-    let values = this.toArray()
-    let output = new Array<string>(values.length)
-    for (let i: i32 = 0; i < values.length; i++) {
-      output[i] = values[i].toString()
-    }
-    return output
-  }
-
-  toI32Array(): Array<i32> {
-    let values = this.toArray()
-    let output = new Array<i32>(values.length)
-    for (let i: i32 = 0; i < values.length; i++) {
-      output[i] = values[i].toI32()
-    }
-    return output
-  }
-
-  toBigIntArray(): Array<BigInt> {
-    let values = this.toArray()
-    let output = new Array<BigInt>(values.length)
-    for (let i: i32 = 0; i < values.length; i++) {
-      output[i] = values[i].toBigInt()
-    }
-    return output
-  }
-
-  toBigDecimalArray(): Array<BigDecimal> {
-    let values = this.toArray()
-    let output = new Array<BigDecimal>(values.length)
-    for (let i: i32 = 0; i < values.length; i++) {
-      output[i] = values[i].toBigDecimal()
-    }
-    return output
-  }
-
-  static fromBooleanArray(input: Array<boolean>): Value {
-    let output = new Array<Value>(input.length)
-    for (let i: i32 = 0; i < input.length; i++) {
-      output[i] = Value.fromBoolean(input[i])
-    }
-    return Value.fromArray(output)
-  }
-
-  static fromBytesArray(input: Array<Bytes>): Value {
-    let output = new Array<Value>(input.length)
-    for (let i: i32 = 0; i < input.length; i++) {
-      output[i] = Value.fromBytes(input[i])
-    }
-    return Value.fromArray(output)
-  }
-
-  static fromI32Array(input: Array<i32>): Value {
-    let output = new Array<Value>(input.length)
-    for (let i: i32 = 0; i < input.length; i++) {
-      output[i] = Value.fromI32(input[i])
-    }
-    return Value.fromArray(output)
-  }
-
-  static fromBigIntArray(input: Array<BigInt>): Value {
-    let output = new Array<Value>(input.length)
-    for (let i: i32 = 0; i < input.length; i++) {
-      output[i] = Value.fromBigInt(input[i])
-    }
-    return Value.fromArray(output)
-  }
-
-  static fromBigDecimalArray(input: Array<BigDecimal>): Value {
-    let output = new Array<Value>(input.length)
-    for (let i: i32 = 0; i < input.length; i++) {
-      output[i] = Value.fromBigDecimal(input[i])
-    }
-    return Value.fromArray(output)
-  }
-
-  static fromStringArray(input: Array<string>): Value {
-    let output = new Array<Value>(input.length)
-    for (let i: i32 = 0; i < input.length; i++) {
-      output[i] = Value.fromString(input[i])
-    }
-    return Value.fromArray(output)
-  }
-
-  static fromArray(input: Array<Value>): Value {
-    let value = new Value()
-    value.kind = ValueKind.ARRAY
-    value.data = input as u64
-    return value
-  }
-
-  static fromBigInt(n: BigInt): Value {
-    let value = new Value()
-    value.kind = ValueKind.BIGINT
-    value.data = n as u64
-    return value
-  }
-
-  static fromBoolean(b: boolean): Value {
-    let value = new Value()
-    value.kind = ValueKind.BOOL
-    value.data = b ? 1 : 0
-    return value
-  }
-
-  static fromBytes(bytes: Bytes): Value {
-    let value = new Value()
-    value.kind = ValueKind.BYTES
-    value.data = bytes as u64
-    return value
-  }
-
-  static fromNull(): Value {
-    let value = new Value()
-    value.kind = ValueKind.NULL
-    return value
-  }
-
-  static fromI32(n: i32): Value {
-    let value = new Value()
-    value.kind = ValueKind.INT
-    value.data = n as u64
-    return value
-  }
-
-  static fromString(s: string): Value {
-    let value = new Value()
-    value.kind = ValueKind.STRING
-    value.data = changetype<u32>(s)
-    return value
-  }
-
-  static fromBigDecimal(n: BigDecimal): Value {
-    let value = new Value()
-    value.kind = ValueKind.BIGDECIMAL
-    value.data = n as u64
-    return value
-  }
-}
-
-/**
- * Common representation for entity data, storing entity attributes
- * as `string` keys and the attribute values as dynamically-typed
- * `Value` objects.
- */
-export class Entity extends TypedMap<string, Value> {
-  unset(key: string): void {
-    this.set(key, Value.fromNull())
-  }
-
-  /** Assigns properties from sources to this Entity in right-to-left order */
-  merge(sources: Array<Entity>): Entity {
-    var target = this
-    for (let i = 0; i < sources.length; i++) {
-      let entries = sources[i].entries
-      for (let j = 0; j < entries.length; j++) {
-        target.set(entries[j].key, entries[j].value)
-      }
-    }
-    return target
-  }
 }
 
 /**
